Refactor ClientManagementPage component to streamline exports
- Changed the export method for the ClientManagementPage component to a default export from its index file.
- This refactor improves the organization of the component structure and simplifies import paths for better maintainability.
